Added Functionoids, Camera movement.
Functionoids can now be used to implement callbacks. As such, the EventHandler
class allows a user to register a functionoid belonging to a class to be
registered with a key. When the key is pressed, the function is called. This is
now implemented for Camera rotate and move, for starters.