Started wrapping SDL
SDL passes the registered keypresses into a small std::set of keys pressed
at that moment. The keys are compared to bound keys for functionoids which are
then called upon.
I only need to allow passing multiple keys, and instead of comparing with one
key, compare with the entire set of keys, so key combo's can be used as well.