Added combination key-support. Added Window Close method.
Keys are now input as a std::set (that also has a typedef for ease of coding),
so the sets can be compared to what keys are input and registered. So far so
good. Also added an Escape key to the enum to allow for window closing. To do
this I've also added a Context::close() method that sets the internal
windowOpened bool to false.