Updated Text class, added Font class.
The new Font class now generates a bitmap font on the fly at creation, after
which the Text class uses the texture data and the texture coordinate data
created for every ASCII character to create strings of quads to be displayed.
Context has been updated to allow for transparent orthographic displaying of
text.